Use state from outside tail to manage seen events
If the client is re-created from the outer main loop, then the Tail
struct will be recreated as well. This means that the state of the seen
events will go down with the ship. We want this state to outlive the
Tail struct, so we pass it in as a parameter and create the state
outside the main loop.