Only create out stream once
This occurs before the loop. Now that we are only locking (which is not
ideal in an async context...) when we print, we don't get deadlocks any
more.

Actually we _are_ in an async context, but the async work occurs before,
and we only have one task active at once.