Use latest time to prevent double printing
We know that the events are guaranteed to come in in reverse
chronological order, so we can optimise - if the current event is older
than what we've already seen then we don't need to continue. This will
help greatly when fetching stacks with long histories - we no longer
have to fetch the entire history of the stack before printing.